backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blessed==1.21.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
ddgs==9.0.0
distro==1.9.0
duckduckgo_search==8.1.1
editor==1.6.6
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-cloud-translate==3.21.1
google-search-results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.29.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
idna==3.10
iniconfig==2.1.0
inquirer==3.4.0
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-groq==0.3.5
langchain-openai==0.3.27
langchain-pinecone==0.2.8
langchain-tests==0.3.20
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langdetect==1.0.9
langsmith==0.4.4
llvmlite==0.44.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
orjson==3.10.18
packaging==24.2
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==3.2.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyuseragents==1.0.5
PyYAML==6.0.2
readchar==4.2.1
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
runs==1.2.2
safeIO==1.2
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==5.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
syrupy==4.9.1
tavily-python==0.7.9
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
translatepy==2.3
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
vcrpy==7.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.2
xmod==1.8.1
yarl==1.20.1
zstandard==0.23.0
